requirements.txt
openai
python-dotenv
PyMuPDF
streamlit